requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
mongoengine==0.27.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dateutil==2.8.2